# openapi_contract.json
{
  "components": {
    "schemas": {
...
      "FeatureFlagSetIn": {
        "properties": {
          "allowlist": {
            "items": {
              "type": "string"
            },
            "title": "Allowlist",
            "type": "array"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "flag_key": {
            "maxLength": 120,
            "minLength": 2,
            "title": "Flag Key",
            "type": "string"
          },
          "rollout_pct": {
            "default": 0,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Rollout Pct",
            "type": "integer"
          }
        },
        "required": [
          "flag_key",
          "enabled"
        ],
        "title": "FeatureFlagSetIn",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NoShowPolicyIn": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "fee_amount": {
            "minimum": 0.0,
            "title": "Fee Amount",
            "type": "number"
          },
          "grace_minutes": {
            "default": 10,
            "maximum": 240.0,
            "minimum": 0.0,
            "title": "Grace Minutes",
            "type": "integer"
          }
        },
        "required": [
          "enabled",
          "fee_amount"
        ],
        "title": "NoShowPolicyIn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Telegram-driven salon management API",
    "title": "SalonOS",
    "version": "1.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/platform/flags": {
      "get": {
        "operationId": "get_feature_flags_api_platform_flags_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-tenant-slug",
            "required": false,
            "schema": {
              "title": "X-Tenant-Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "object"
                  },
                  "title": "Response Get Feature Flags Api Platform Flags Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Feature Flags",
        "tags": [
          "platform"
        ]
      },
      "put": {
        "operationId": "set_feature_flag_api_platform_flags_put",
        "parameters": [
          {
            "in": "header",
            "name": "x-tenant-slug",
            "required": false,
            "schema": {
              "title": "X-Tenant-Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeatureFlagSetIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Set Feature Flag Api Platform Flags Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Feature Flag",
        "tags": [
          "platform"
        ]
      }
    },
    "/api/platform/flags/evaluate": {
      "get": {
        "operationId": "evaluate_feature_flag_api_platform_flags_evaluate_get",
        "parameters": [
          {
            "in": "query",
            "name": "flag_key",
            "required": true,
            "schema": {
              "maxLength": 120,
              "minLength": 2,
              "title": "Flag Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "subject_key",
            "required": false,
            "schema": {
              "title": "Subject Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-slug",
            "required": false,
            "schema": {
              "title": "X-Tenant-Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Evaluate Feature Flag Api Platform Flags Evaluate Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evaluate Feature Flag",
        "tags": [
          "platform"
        ]
      }
    },
    "/api/platform/idempotency/cleanup": {
      "post": {
        "operationId": "idempotency_cleanup_api_platform_idempotency_cleanup_post",
        "parameters": [
          {
            "in": "query",
            "name": "older_than_hours",
            "required": false,
            "schema": {
              "default": 168,
              "maximum": 8760.0,
              "minimum": 1.0,
              "title": "Older Than Hours",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-slug",
            "required": false,
            "schema": {
              "title": "X-Tenant-Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Idempotency Cleanup Api Platform Idempotency Cleanup Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Idempotency Cleanup",
        "tags": [
          "platform"
        ]
      }
    },
...
    "/api/platform/no-show-policy": {
      "get": {
        "operationId": "get_no_show_policy_api_platform_no_show_policy_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-tenant-slug",
            "required": false,
            "schema": {
              "title": "X-Tenant-Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get No Show Policy Api Platform No Show Policy Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get No Show Policy",
        "tags": [
          "platform"
        ]
      },
      "put": {
        "operationId": "set_no_show_policy_api_platform_no_show_policy_put",
        "parameters": [
          {
            "in": "header",
            "name": "x-tenant-slug",
            "required": false,
            "schema": {
              "title": "X-Tenant-Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoShowPolicyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Set No Show Policy Api Platform No Show Policy Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set No Show Policy",
        "tags": [
          "platform"
        ]
      }
    },
    "/api/platform/outbox/cleanup": {
      "post": {
        "operationId": "cleanup_outbox_api_platform_outbox_cleanup_post",
        "parameters": [
          {
            "in": "query",
            "name": "older_than_hours",
            "required": false,
            "schema": {
              "default": 168,
              "maximum": 8760.0,
              "minimum": 1.0,
              "title": "Older Than Hours",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-slug",
            "required": false,
            "schema": {
              "title": "X-Tenant-Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Cleanup Outbox Api Platform Outbox Cleanup Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cleanup Outbox",
        "tags": [
          "platform"
        ]
      }
    },
    "/api/platform/outbox/dispatch": {
      "post": {
        "operationId": "dispatch_outbox_api_platform_outbox_dispatch_post",
        "parameters": [
          {
            "in": "query",
            "name": "batch_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500.0,
              "minimum": 1.0,
              "title": "Batch Size",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-slug",
            "required": false,
            "schema": {
              "title": "X-Tenant-Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Dispatch Outbox Api Platform Outbox Dispatch Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dispatch Outbox",
        "tags": [
          "platform"
        ]
      }
    },
    "/api/platform/outbox/events": {
      "get": {
        "operationId": "get_outbox_events_api_platform_outbox_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 1000.0,
              "minimum": 1.0,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-slug",
            "required": false,
            "schema": {
              "title": "X-Tenant-Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "object"
                  },
                  "title": "Response Get Outbox Events Api Platform Outbox Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Outbox Events",
        "tags": [
          "platform"
        ]
      }
    },
...
    "/api/platform/outbox/retry-failed": {
      "post": {
        "operationId": "retry_failed_outbox_api_platform_outbox_retry_failed_post",
        "parameters": [
          {
            "in": "query",
            "name": "include_dead_letter",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Dead Letter",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000.0,
              "minimum": 1.0,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-slug",
            "required": false,
            "schema": {
              "title": "X-Tenant-Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Retry Failed Outbox Api Platform Outbox Retry Failed Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Failed Outbox",
        "tags": [
          "platform"
        ]
      }
    },
    "/api/platform/payments/intents": {
      "get": {
        "operationId": "list_payment_intents_endpoint_api_platform_payments_intents_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 1000.0,
              "minimum": 1.0,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-slug",
            "required": false,
            "schema": {
              "title": "X-Tenant-Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "object"
                  },
                  "title": "Response List Payment Intents Endpoint Api Platform Payments Intents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Payment Intents Endpoint",
        "tags": [
          "platform"
        ]
      },
      "post": {
        "operationId": "create_payment_intent_endpoint_api_platform_payments_intents_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-tenant-slug",
            "required": false,
            "schema": {
              "title": "X-Tenant-Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentIntentIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Create Payment Intent Endpoint Api Platform Payments Intents Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Payment Intent Endpoint",
        "tags": [
          "platform"
        ]
      }
    },
    "/api/platform/payments/intents/{payment_intent_id}/capture": {
      "post": {
        "operationId": "capture_payment_intent_endpoint_api_platform_payments_intents__payment_intent_id__capture_post",
        "parameters": [
          {
            "in": "path",
            "name": "payment_intent_id",
            "required": true,
            "schema": {
              "title": "Payment Intent Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "provider_ref",
            "required": false,
            "schema": {
              "title": "Provider Ref",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-slug",
            "required": false,
            "schema": {
              "title": "X-Tenant-Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Capture Payment Intent Endpoint Api Platform Payments Intents  Payment Intent Id  Capture Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Capture Payment Intent Endpoint",
        "tags": [
          "platform"
        ]
      }
    },
...
  }
}